Creating separate header for addition kernels
